+++ src/homematicip/class_maps.py
@@ -91,6 +91,7 @@ TYPE_CLASS_MAP = {
     DeviceType.SMOKE_DETECTOR: SmokeDetector,
     DeviceType.SOIL_MOISTURE_SENSOR_INTERFACE: SoilMoistureSensorInterface,
     DeviceType.STATUS_BOARD_8: StatusBoard8,
+    DeviceType.SWITCH_MEASURING_CABLE_INDOOR: SwitchMeasuring,
     DeviceType.SWITCH_MEASURING_CABLE_OUTDOOR: SwitchMeasuring,
     DeviceType.TEMPERATURE_HUMIDITY_SENSOR: TemperatureHumiditySensorWithoutDisplay,
     DeviceType.TEMPERATURE_HUMIDITY_SENSOR_COMPACT: TemperatureHumiditySensorOutdoor,